Fix Windows x64 build broken by r5299. Review URL: http://codereview.chromium.org/3158019
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@5300 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -390,7 +390,7 @@ void MemoryAllocator::FreeRawMemory(void* mem,
   }
   Counters::memory_allocated.Decrement(static_cast<int>(length));
   size_ -= static_cast<int>(length);
-  if (executable == EXECUTABLE) size_executable_ -= length;
+  if (executable == EXECUTABLE) size_executable_ -= static_cast<int>(length);
   ASSERT(size_ >= 0);
 }
